refactor: split peers.rs (3936 lines) into sub-module directory
peers.rs → peers/
  mod.rs         (~258 lines) - types, config, path helpers, re-exports
  heartbeat.rs   (~340 lines) - heartbeat CRUD, coordination event writers
  board.rs       (~240 lines) - compute_board_state, compaction
  discovery.rs   (~200 lines) - discover_active_peers, discover_all_sessions, infer_session_id
  render_coord.rs (~332 lines) - coordination protocol + peer updates rendering
  autoclaim.rs   (~230 lines) - auto-claim logic, scope derivation
  helpers.rs     (~155 lines) - format_age, truncate_to_budget, parse_rfc3339
  tests.rs       (~2224 lines) - all existing tests preserved unchanged

All public API unchanged (lib.rs re-exports unaffected).
Part of GH-209.
